Docker label for new rest api
When a model runner implementation is following the micro runner standards https://wiki.acumos.org/pages/viewpage.action?pageId=17696349 we need to have nginx proxy change to satify this new standard. It is possible that there are multiple docker images built with the older model runner standards which will not have this label.

Issue-ID: ACUMOS-2782

Change-Id: Ife60e6b793cb65f6055274fc6679adc0e8e96683
Signed-off-by: justin14 <justin.early@est.tech>