Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Looks OK. I can +2/merge if you like,
> 
> But what is the criterion for which packages get moved to onboarding-base-r? Why not move them all?

At first I would like to do this and so remove the packages.R file from on-boarding repo. But I think that the onboarding server used the packages.R file to write the dependencies listed in metadata.json and added these dependencies in packages.R file. So if this file doesn't exist what happens ? 

So for the moment I only remove the common packages that are used by all the R models whatever the algorithms librarie used (like RProtoBuf for example). But if you confirmed that I can completly remove packages.R file and move all these intall.packages() commands to micro-service repos, I will do that

Patch-set: 1